pyproject.toml
[build-system]
requires = ["hatchling>=1.21,<2.0"]
build-backend = "hatchling.build"

[project]
name = "claudehooks"
version = "0.1.0"
description = "Add Claude as middleware to any Python app with a single decorator"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "MavPro Group LLC" },
]
keywords = ["claude", "anthropic", "llm", "middleware", "decorator", "ai", "claudehooks"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "anthropic>=0.40.0,<2.0",
    "pydantic>=2.0.0,<3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
    "fastapi>=0.100",
    "httpx>=0.24",
    "openai>=1.30",
]
fastapi = [
    "fastapi>=0.100,<1.0",
    "httpx>=0.24,<1.0",
]
openai = [
    "openai>=1.30,<3.0",
]
pii = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
]
telemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.urls]
Homepage = "https://github.com/mavprogroup/claudehooks"
Documentation = "https://github.com/mavprogroup/claudehooks#readme"
Repository = "https://github.com/mavprogroup/claudehooks"
Issues = "https://github.com/mavprogroup/claudehooks/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/claudehooks"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100
